// x-pack/solutions/observability/plugins/slo/server/services/transform_generators/synthetics_availability.d.ts
import type { TransformPutTransformRequest } from '@elastic/elasticsearch/lib/api/types';
import type { DataViewsService } from '@kbn/data-views-plugin/common';
import { TransformGenerator } from '.';
import type { SLODefinition } from '../../domain/models';
export declare class SyntheticsAvailabilityTransformGenerator extends TransformGenerator {
    constructor(spaceId: string, dataViewService: DataViewsService, isServerless: boolean);
    getTransformParams(slo: SLODefinition): Promise<TransformPutTransformRequest>;
    private buildTransformId;
    private buildGroupBy;
    private buildSource;
    private buildDestination;
    private buildAggregations;
}
export declare const buildParamValues: (params: Record<string, Array<{
    label: string;
    value: string;
}>>) => Record<string, string[]>;